+++ NEWS.md
@@ -1,6 +1,6 @@
-# yulab.utils 0.0.6.004
+# yulab.utils 0.0.6.005
 
-+ `read_with_cache()` and `read_tsv_with_cache()` (2023-08-02, Wed) 
++ `yread()` and `yread_tsv()` that read file with caching (2023-08-02, Wed) 
 + `rbindlist()` to rbind a list (2023-08-01, Tue)
 + `exec()` to run system command (2023-06-19, Mon)
 + `scale_range()` to normalize data by range (2023-06-18, Sun)